Archives/Dockerfiles/Spark/requirements.txt
confluent-kafka
protobuf
delta-spark==3.1.0
gtfs-realtime-bindings